+++ src/mapfile_parser/frontends/pj64_syms.py
@@ -6,6 +6,7 @@
 from __future__ import annotations
 
 import argparse
+from collections.abc import Callable
 import decomp_settings
 from pathlib import Path
 from typing import TextIO
@@ -21,12 +22,18 @@ def writePj64SymsToFile(mapFile: mapfile.MapFile, outFile: TextIO):
                 symType = "code" if file.sectionType == ".text" else "data"
                 outFile.write(f"{sym.vram:08X},{symType},{sym.name}\n")
 
-def doPj64Syms(mapPath: Path, outputPath: Path|None) -> int:
+def doPj64Syms(
+        mapPath: Path,
+        outputPath: Path|None,
+        plfResolver: Callable[[Path], Path|None]|None=None,
+    ) -> int:
     if not mapPath.exists():
         print(f"Could not find mapfile at '{mapPath}'")
         return 1
 
     mapFile = mapfile.MapFile.newFromMapFile(mapPath)
+    if plfResolver is not None:
+        mapFile = mapFile.resolvePartiallyLinkedFiles(plfResolver)
 
     if outputPath is None:
         writePj64SymsToFile(mapFile, sys.stdout)
@@ -47,7 +54,20 @@ def processArguments(args: argparse.Namespace, decompConfig: decomp_settings.Con
 
     outputPath: Path = args.output
 
-    exit(doPj64Syms(mapPath, outputPath))
+    plfExt: list[str]|None = args.plf_ext
+
+    plfResolver = None
+    if plfExt is not None:
+        def resolver(x: Path) -> Path|None:
+            if x.suffix in plfExt:
+                newPath = x.with_suffix(".map")
+                if newPath.exists():
+                    return newPath
+            return None
+
+        plfResolver = resolver
+
+    exit(doPj64Syms(mapPath, outputPath, plfResolver=plfResolver))
 
 def addSubparser(subparser: argparse._SubParsersAction[argparse.ArgumentParser], decompConfig: decomp_settings.Config|None=None):
     parser = subparser.add_parser("pj64_syms", help="Produce a PJ64 compatible symbol map.")
@@ -66,5 +86,7 @@ def addSubparser(subparser: argparse._SubParsersAction[argparse.ArgumentParser],
         parser.add_argument("mapfile", help="Path to a map file.", type=Path)
 
     parser.add_argument("output", help="Path to output file. If omitted then output will be written to stdout", type=Path, nargs="?")
+
+    parser.add_argument("-x", "--plf-ext", help="File extension for partially linked files (plf). Will be used to transform the `plf`s path into a mapfile path by replacing the extension. The extension must contain the leading period. This argument can be passed multiple times.", action="append")
 
     parser.set_defaults(func=processArguments)